build(examples): add common ci serve run script
