+++ projects/ngx-meta/examples/src/add-ci-run-scripts.ts
@@ -1,0 +1,43 @@
+import { jsonToString, Log } from './utils.js'
+import { join } from 'path'
+import { PACKAGE_JSON } from './constants.js'
+import { readFile, writeFile } from 'fs/promises'
+
+const BUILD_SSR_RUN_SCRIPT = 'build:ssr'
+//👇 Since v17, this script name includes :{projectName} at the end
+const SERVE_SSR_RUN_SCRIPT_PREFIX = 'serve:ssr'
+const CI_BUILD_RUN_SCRIPT = 'ci:build'
+const CI_SERVE_RUN_SCRIPT = 'ci:serve'
+const BUILD_WITH_SOURCE_MAPS = 'ng build --source-map'
+
+export async function addCiRunScripts(opts: {
+  appDir: string
+  appName: string
+}) {
+  // Builds with SSR + source maps
+  Log.step('Adding build and serve common run scripts for CI/CD')
+  const appPkgJsonFile = join(opts.appDir, PACKAGE_JSON)
+  const appPkgJson = JSON.parse(
+    await readFile(join(opts.appDir, PACKAGE_JSON), 'utf8'),
+  ) as {
+    scripts: Record<string, string>
+  }
+
+  // Build script
+  if (BUILD_SSR_RUN_SCRIPT in appPkgJson.scripts) {
+    appPkgJson.scripts[CI_BUILD_RUN_SCRIPT] =
+      `${BUILD_WITH_SOURCE_MAPS} && ng run ${opts.appName}:server`
+  } else {
+    appPkgJson.scripts[CI_BUILD_RUN_SCRIPT] = BUILD_WITH_SOURCE_MAPS
+  }
+
+  // Serve script
+  const serveSsrRunScript = Object.keys(appPkgJson.scripts).find((runScript) =>
+    runScript.startsWith(SERVE_SSR_RUN_SCRIPT_PREFIX),
+  )
+  if (!serveSsrRunScript) {
+    throw new Error('Cannot find SSR run script for app')
+  }
+  appPkgJson.scripts[CI_SERVE_RUN_SCRIPT] = `pnpm run ${serveSsrRunScript}`
+  await writeFile(appPkgJsonFile, jsonToString(appPkgJson))
+}
