+++ projects/ngx-meta/examples/src/copy-app-dir-into-project.ts
@@ -3,7 +3,7 @@ import { basename, join } from 'path'
 import { cp } from 'fs/promises'
 
 export async function copyAppDirIntoProject(appDir: string) {
-  Log.step('Copying app from into project')
+  Log.step('Copying example app into project')
   const appDirName = basename(appDir)
   const destination = join(getAppsDir(), appDirName)
   await cp(appDir, destination, { recursive: true })
